Only update migration map when Go button is clicked.
Change the approach for updating the migration-map to be more dynamic.
Use onChange in the migration-map and only update when the Go button is
pressed.

This almost works. It works for all the angular bits, but it doesn’t
seem to force the google-chart to redraw. Still need to figure that out.