Fix dynamic chart updating (sort of)
I discovered (accidentally) that if the Home page has a chart, even
though it won’t update, if a linked page does have a chart, it will
dynamically update.

If I remove the chart from the home page, the linked chart no longer
dynamically updates.

So, for now, the migration form and chart are on a linked page
(migrate). When changing code on that page, it is necessary to first go
to the home page and refresh, then click the migrate link.